test8/rag_backend/requirements.txt
streamlit>=1.32.0
PyPDF2>=3.0.1
requests>=2.31.0
numpy>=1.26.4
sentence-transformers>=2.2.2
langchain>=0.1.16
langchain-docling>=0.1.3
python-dotenv>=1.0.0